Lab3/requirements.txt
langchain
langchain-core
langchain-google-genai
python-dotenv
pandas
numpy
scikit-learn
